Remove block_until_head_timeout_eager_prefetch test override
Web platform tests should, in general, run under the same configuration
we ship to our users, or at least the same as experimental web platform
features. This is the first step in removing some of the overrides we
are currently using.

Eventually, it might be good to remove these block_until_head_timeout
parameters, at least for speculation rules.

Bug: 356622798
Change-Id: If4b00cfbfbe5d6c7cf2dd55a3428a129e40868f6